backend/requirements.txt
paho-mqtt
python-dotenv
flask
flask-cors
uvicorn